move shared state logic to new shared.py
